requirements.txt
Flask==3.0.0
APScheduler==3.10.4
SQLAlchemy==2.0.23
requests==2.31.0
PyYAML==6.0.1
python-dateutil==2.8.2
pytz==2023.3
